GH-5064: added docker makefile for Tomcat 9 / Jetty 9 and updated e2e tests

+++ docker/build.sh
@@ -5,45 +5,48 @@ set -e
 CURRENT=$(pwd)
 cd ..
 PROJECT_ROOT=$(pwd)
-cd "$CURRENT"
-
-# clean "ignore" directory
-cd ignore
-rm -f *.zip
-
-cd "$PROJECT_ROOT"
-
-# remove assembly/target since this is not removed by mvn clean
-rm -rf assembly/target/
-
 MVN_VERSION=$(xmllint --xpath "//*[local-name()='project']/*[local-name()='version']/text()" pom.xml)
 
-#Clean, format and package
-echo "Building with Maven"
-mvn clean
-mvn -T 2C formatter:format impsort:sort && mvn xml-format:xml-format
-mvn install -DskipTests
-mvn -Passembly package -DskipTests -Dmaven.javadoc.skip=true -Dformatter.skip=true -Dimpsort.skip=true -Dxml-format.skip=true  -Djapicmp.skip -Denforcer.skip=true -Dbuildnumber.plugin.phase=none -Danimal.sniffer.skip=true
-
-# find .zip file
-ZIP=$(find assembly/target/*.zip)
-echo "$ZIP"
-
-# copy zip file into rdf4j.zip
-cp "$ZIP" "${CURRENT}/ignore/rdf4j.zip"
-
 cd "$CURRENT"
 
-# build
-echo "Building docker image"
-docker compose build --pull --no-cache
+if [ -z ${SKIP_BUILD+x} ]; then
+  # clean "ignore" directory
+  cd ignore
+  rm -f *.zip
 
-docker tag docker-rdf4j:latest eclipse/rdf4j-workbench:${MVN_VERSION}
+  cd "$PROJECT_ROOT"
+
+  # remove assembly/target since this is not removed by mvn clean
+  rm -rf assembly/target/
+
+  #Clean, format and package
+  echo "Building with Maven"
+  mvn clean
+  mvn -T 2C formatter:format impsort:sort && mvn xml-format:xml-format
+  mvn install -DskipTests
+  mvn -Passembly package -DskipTests -Dmaven.javadoc.skip=true -Dformatter.skip=true -Dimpsort.skip=true -Dxml-format.skip=true  -Djapicmp.skip -Denforcer.skip=true -Dbuildnumber.plugin.phase=none -Danimal.sniffer.skip=true
+
+  # find .zip file
+  ZIP=$(find assembly/target/*.zip)
+  echo "$ZIP"
+
+  # copy zip file into rdf4j.zip
+  cp "$ZIP" "${CURRENT}/ignore/rdf4j.zip"
+
+  cd "$CURRENT"
+fi
+
+# build
+APP_SERVER=${APP_SERVER:-tomcat}
+echo "Building docker image for ${APP_SERVER}"
+
+docker compose build --pull --no-cache
+docker tag docker-rdf4j:latest eclipse/rdf4j-workbench-${APP_SERVER}:${MVN_VERSION}
 
 echo "
 Docker image tagged as:
-  docker-rdf4j:latest
-  eclipse/rdf4j-workbench:${MVN_VERSION}
+    docker-rdf4j:latest
+    eclipse/rdf4j-workbench-${APP_SERVER}:${MVN_VERSION}
 
 To start the workbench and server:
     docker compose up -d
